Create Pre-In-Post-Order_Traversal.cpp
On the first line, there will be a single number, N, showing the number of nodes in the tree. On the next N lines, the ith line contains the data and the left and right child of the ith node, for example if the 5th of the N lines reads abc 4 6 it means that the node number 5 contains the string “abc”, its left child is the node number 4 and its right child is the node number 6. If the node does not have a child, or both, it would be shown by a -1 in the input. In other words, a leaf containing the string “aa” would be written as aa -1 -1 in the input. The node numbers are 0-based and the 0th node is the root of the tree. For example take a look at the tree below. The numbers below the nodes show the corresponding line of input for the nodes.

Input Format:
9 
a 1 3 
b 2 8
d 6 7 
c -1 5 
i -1 -1 
f 4 -1 
g -1 -1 
h -1 -1 
e -1 -1